third_party/WebKit/LayoutTests/fast/dom/navigator-vibration-expected.txt

Test for the Vibration API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.vibrate(0); is true
PASS navigator.vibrate([]); is true
PASS navigator.vibrate(1); is true
PASS navigator.vibrate([1, 2]); is true
PASS navigator.vibrate([1, 2, 3]); is true
PASS navigator.vibrate(10000); is true
PASS navigator.vibrate(10001); is false
PASS navigator.vibrate(-1); is false
PASS var p = []; for (var i = 0; i < 99; i++) p[i] = 1; navigator.vibrate(p); is true
PASS var p = []; for (var i = 0; i < 100; i++) p[i] = 1; navigator.vibrate(p); is false
PASS navigator.vibrate(); threw exception TypeError: Failed to execute 'vibrate' on 'Navigator': 1 argument required, but only 0 present..
PASS navigator.vibrate([1], [2]); threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
